fixed refund job, implemented mixin for offer controller methods, refactoring
