+++ app/controllers/english_offers_controller.rb
@@ -1,15 +1,11 @@
 # rubocop:disable Metrics
 class EnglishOffersController < ApplicationController
   include BeforeRender
+  include Offerable
   protect_from_forgery with: :null_session
 
-  before_action :authenticate_user!
-  before_action :find_auction, only: %i[new create]
-  before_action :check_for_ban, only: :create
   before_action :set_offer, only: %i[show edit update]
   before_render :find_or_initialize_autobidder, only: %i[new create edit update]
-  before_action :authorize_phone_confirmation
-  # before_action :authorize_offer_for_user, except: %i[new create]
   before_action :prevent_check_for_invalid_bid, only: [:update]
 
   include RecaptchaValidatable
@@ -18,13 +14,7 @@ class EnglishOffersController < ApplicationController
 
   # GET /auctions/aa450f1a-45e2-4f22-b2c3-f5f46b5f906b/offers/new
   def new
-    offer = @auction.offer_from_user(current_user.id)
-
-    if turbo_frame_request?
-      render turbo_stream: turbo_stream.action(:redirect, root_path), notice: t('offers.already_exists') and return
-    else
-      redirect_to root_path, status: :see_other, notice: t('offers.already_exists') and return
-    end if offer
+    prevent_check_for_existed_offer and return if @auction.offer_from_user(current_user.id)
 
     BillingProfile.create_default_for_user(current_user.id)
     @offer = Offer.new(auction_id: @auction.id, user_id: current_user.id)
@@ -50,7 +40,7 @@ class EnglishOffersController < ApplicationController
     if recaptcha_valid
       if create_predicate(@auction)
         broadcast_update_auction_offer(@auction)
-        send_outbided_notification(auction: @auction, offer: @offer, flash: flash)
+        send_outbided_notification(auction: @auction, offer: @offer, flash:)
         update_auction_values(@auction, t('english_offers.create.created'))
       else
         flash[:alert] = if @offer.errors.full_messages_for(:cents).present?
@@ -88,7 +78,7 @@ class EnglishOffersController < ApplicationController
     if recaptcha_valid
       if update_predicate(@auction)
         broadcast_update_auction_offer(@auction)
-        send_outbided_notification(auction: @auction, offer: @offer, flash: flash)
+        send_outbided_notification(auction: @auction, offer: @offer, flash:)
         update_auction_values(@auction, t('english_offers.edit.bid_updated'))
       else
         flash[:alert] = if @offer.errors.full_messages_for(:cents).present?
@@ -100,7 +90,6 @@ class EnglishOffersController < ApplicationController
         redirect_to root_path
       end
     else
-
       @show_checkbox_recaptcha = true unless @success
       flash.now[:alert] = t('english_offers.form.captcha_verification')
       redirect_to root_path, status: :see_other
@@ -109,26 +98,12 @@ class EnglishOffersController < ApplicationController
 
   private
 
-  def find_auction
-    @auction = Auction.english.find_by!(uuid: params[:auction_uuid], platform: 'english')
-  end
-
   def find_or_initialize_autobidder
     @autobider = current_user&.autobiders&.find_or_initialize_by(domain_name: @auction.domain_name)
   end
 
-  def update_not_allowed(auction)
-    !auction.english? || !auction.in_progress?
-  end
-
-  def check_for_ban
-    if Ban.valid.where(user_id: current_user).where(domain_name: @auction.domain_name).any? || current_user.completely_banned?
-      redirect_to root_path, flash: { alert: I18n.t('.english_offers.create.ban') } and return
-    end
-  end
-
   def broadcast_update_auction_offer(auction)
-    Offers::UpdateBroadcastService.call({ auction: auction })
+    Offers::UpdateBroadcastService.call({ auction: })
   end
 
   def update_auction_values(auction, message_text)
@@ -142,9 +117,10 @@ class EnglishOffersController < ApplicationController
   def prevent_check_for_invalid_bid
     auction = Auction.with_user_offers(current_user.id).find_by(uuid: @offer.auction.uuid)
 
-    return unless bid_is_bad?(auction: auction, update_params: update_params)
+    return unless bid_is_bad?(auction:, update_params:)
 
-    flash[:alert] = "#{t('english_offers.show.bid_failed', price: format('%.2f', auction.highest_price.to_f).tr('.', ','))}"
+    flash[:alert] =
+      "#{t('english_offers.show.bid_failed', price: format('%.2f', auction.highest_price.to_f).tr('.', ','))}"
 
     if turbo_frame_request?
       render turbo_stream: turbo_stream.action(:redirect, root_path)
@@ -199,25 +175,5 @@ class EnglishOffersController < ApplicationController
   def update_params
     update_params = params.require(:offer).permit(:price, :billing_profile_id)
     merge_updated_by(update_params)
-  end
-
-  def set_offer
-    @offer = current_user.offers.find_by!(uuid: params[:uuid])
-  end
-
-  def authorize_phone_confirmation
-    return unless current_user.requires_phone_number_confirmation?
-
-    flash[:notice] = t('phone_confirmations.confirmation_required')
-
-    if turbo_frame_request?
-      render turbo_stream: turbo_stream.action(:redirect, new_user_phone_confirmation_path(current_user.uuid))
-    else
-      redirect_to new_user_phone_confirmation_path(current_user.uuid), status: :see_other
-    end
-  end
-
-  def authorize_offer_for_user
-    authorize! :manage, @offer
   end
 end
